add Authentication  with Firebase

+++ src/common/components/Task/Task.tsx
@@ -1,13 +1,9 @@
 import s from "./Task.module.css";
 import { EditableTitle } from "../EditableTitle/EditableTitle";
-import { useAppDispatch } from "../../../redux/store";
+import { useAppDispatch, useAppSelector } from "../../../redux/store";
 import { Button } from "../Button/Button";
 import { StatusTask } from "../../../redux/taskSlice";
-import {
-  changeTaskStatusAsync,
-  removeTaskAsync,
-  updateTaskTitleAsync,
-} from "../../../redux/taskThunk";
+import { changeTaskStatusAsync, removeTaskAsync, updateTaskTitleAsync } from "../../../redux/taskThunk";
 
 type TaskProps = {
   taskName: string;
@@ -16,39 +12,32 @@ type TaskProps = {
   taskStatus: StatusTask;
 };
 
-export const Task = ({
-  taskName,
-  todolistId,
-  taskId,
-  taskStatus,
-}: TaskProps) => {
+export const Task = ({ taskName, todolistId, taskId, taskStatus }: TaskProps) => {
   const dispatch = useAppDispatch();
+  const uid = useAppSelector((state) => state.auth.uid);
+
+  if (!uid) {
+    return <div>Unauthorized</div>; // Если uid нет, то показываем сообщение
+  }
 
   const handleDeleteTask = () => {
-    dispatch(removeTaskAsync({ todolistId, taskId }));
+    dispatch(removeTaskAsync({ uid, todolistId, taskId }));
   };
 
   const handleChangeTaskStatus = () => {
     const newStatus = taskStatus === "active" ? "complete" : "active";
-    dispatch(changeTaskStatusAsync({ todolistId, taskId, status: newStatus }));
+    dispatch(changeTaskStatusAsync({ uid, todolistId, taskId, status: newStatus }));
   };
 
   const handleChangeTaskTitle = (newTitle: string) => {
-    dispatch(updateTaskTitleAsync({ todolistId, taskId, title: newTitle }));
+    dispatch(updateTaskTitleAsync({ uid, todolistId, taskId, title: newTitle }));
   };
 
   return (
     <div className={s.taskBlock}>
       <div className={s.taskContent}>
-        <input
-          type="checkbox"
-          checked={taskStatus === "complete"}
-          onChange={handleChangeTaskStatus}
-        />
-        <EditableTitle
-          title={taskName}
-          onChange={handleChangeTaskTitle}
-        ></EditableTitle>
+        <input type="checkbox" checked={taskStatus === "complete"} onChange={handleChangeTaskStatus} />
+        <EditableTitle title={taskName} onChange={handleChangeTaskTitle}></EditableTitle>
       </div>
 
       <Button useIcon={true} onClick={handleDeleteTask} type="delete" />
